+++ core/shared/platform/common/posix/posix_memmap.c
@@ -14,7 +14,7 @@ os_mmap(void *hint, size_t size, int prot, int flags)
     uint8 *addr;
     uint32 i;
 
-    page_size = getpagesize();
+    page_size = (uint64)getpagesize();
     request_size = (size + page_size - 1) & ~(page_size - 1);
 
     if ((size_t)request_size < size)
@@ -60,7 +60,7 @@ os_mmap(void *hint, size_t size, int prot, int flags)
 void
 os_munmap(void *addr, size_t size)
 {
-    uint64 page_size = getpagesize();
+    uint64 page_size = (uint64)getpagesize();
     uint64 request_size = (size + page_size - 1) & ~(page_size - 1);
 
     if (addr) {
@@ -75,7 +75,7 @@ int
 os_mprotect(void *addr, size_t size, int prot)
 {
     int map_prot = PROT_NONE;
-    uint64 page_size = getpagesize();
+    uint64 page_size = (uint64)getpagesize();
     uint64 request_size = (size + page_size - 1) & ~(page_size - 1);
 
     if (!addr)
